refactor: reorganize horizontal table widget
Split horizontal table content and style controls into modular files
Update widget manager config to support widget_file and control_dirs
Move widget file into horizontal-table directory and clean old path
