+++ specs/001-split-widget-controls/tasks.md
@@ -21,7 +21,7 @@ description: "Task list for separating widget controls only"
 ## Path Conventions
 
 - **Plugin root**: `saaspricing.php`, `includes/`, `assets/`, `languages/`
-- **Widget entry files**: `includes/widgets/comparison-table/saaspricing-comparison-table.php`, `includes/widgets/saaspricing-horizontal-table.php`, `includes/widgets/saaspricing-pricelist.php`, `includes/widgets/saaspricing-vertical-table.php`
+- **Widget entry files**: `includes/widgets/comparison-table/saaspricing-comparison-table.php`, `includes/widgets/horizontal-table/saaspricing-horizontal-table.php`, `includes/widgets/saaspricing-pricelist.php`, `includes/widgets/saaspricing-vertical-table.php`
 - **Control files**: `includes/widgets/<widget>/controls/content-controls-manager.php`,
   `includes/widgets/<widget>/controls/style-controls-manager.php`, and tab-prefixed section files
 - Every move task MUST preserve an old-to-new mapping for the touched control areas
@@ -62,7 +62,7 @@ description: "Task list for separating widget controls only"
 ### Implementation for User Story 1
 
 - [X] T006 [P] [US1] Update `includes/widgets/comparison-table/saaspricing-comparison-table.php` so it keeps widget logic in place, uses local comparison traits directly, and does not `require_once` raw control snippets inline
-- [ ] T007 [P] [US1] Update `includes/widgets/saaspricing-horizontal-table.php` so it keeps widget logic in place, uses horizontal traits directly, and does not `require_once` its trait files inline
+- [X] T007 [P] [US1] Update `includes/widgets/horizontal-table/saaspricing-horizontal-table.php` so it keeps widget logic in place, uses local horizontal traits directly, and does not `require_once` raw control snippets inline
 - [ ] T008 [P] [US1] Update `includes/widgets/saaspricing-pricelist.php` so it keeps widget logic in place, uses pricelist traits directly, and does not `require_once` its trait files inline
 - [ ] T009 [P] [US1] Update `includes/widgets/saaspricing-vertical-table.php` so it keeps widget logic in place, uses vertical traits directly, and does not `require_once` its trait files inline
 - [ ] T010 [US1] Verify the current widget files and registration flow remain unchanged after control extraction
@@ -85,7 +85,7 @@ description: "Task list for separating widget controls only"
 
 - [X] T012 [P] [US2] Extract comparison content-tab controls from `includes/widgets/comparison-table/saaspricing-comparison-table.php` into `includes/widgets/comparison-table/controls/content-controls-manager.php` and any required `includes/widgets/comparison-table/controls/content-controls/content-*.php` section traits
 - [X] T013 [P] [US2] Extract comparison style-tab controls from `includes/widgets/comparison-table/saaspricing-comparison-table.php` into `includes/widgets/comparison-table/controls/style-controls-manager.php` and any required `includes/widgets/comparison-table/controls/style-controls/style-*.php` section traits
-- [ ] T014 [P] [US2] Extract horizontal content-tab and style-tab controls into a widget-owned folder with `controls/content-controls-manager.php`, `controls/style-controls-manager.php`, and any required tab-prefixed section traits
+- [X] T014 [P] [US2] Extract horizontal content-tab and style-tab controls into `includes/widgets/horizontal-table/controls/content-controls-manager.php`, `includes/widgets/horizontal-table/controls/style-controls-manager.php`, and the required tab-prefixed section traits
 - [ ] T015 [P] [US2] Extract pricelist content-tab and style-tab controls into a widget-owned folder with `controls/content-controls-manager.php`, `controls/style-controls-manager.php`, and any required tab-prefixed section traits
 - [ ] T016 [P] [US2] Extract vertical content-tab and style-tab controls into a widget-owned folder with `controls/content-controls-manager.php`, `controls/style-controls-manager.php`, and any required tab-prefixed section traits
 - [ ] T017 [US2] Verify every widget still has dedicated content and style control manager files even when a tab has only one section
